requirements.txt
# general modules
django >3.2, <3.3
python-decouple
pillow
requests
beautifulsoup4

# databases
psycopg2-binary

# refactor & test modules
Faker
flake8
black

# third party
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
djangorestframework-jwt
drf-yasg[validation]
django-mail-templated

# deployment

# Background processing and cache
